게시물 삭제시 권한체크

+++ src/main/java/com/taeate/example/demo/controller/UsrArticleController.java
@@ -53,12 +53,29 @@ public class UsrArticleController {
 
     @RequestMapping("/usr/article/doDelete")
     @ResponseBody
-    public ResultData doDelete(int id) {
+    public ResultData doDelete(HttpSession httpSession, int id) {
+        boolean isLogined = false;
+        int loginedMemberId = 0;
+        
+
+        if (httpSession.getAttribute("loginedMemberId") != null){
+            isLogined = true;
+            loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+        }
+
+        if (isLogined == false) {
+            return ResultData.from("F-A", "로그인 후 이용해주세요.");
+        }
+
         Article article = articleService.getArticle(id);
 
+        if (article.getMemberId() != loginedMemberId) {
+            return ResultData.from("F-2", "권한이 없습니다.");
+        }
         if (article == null) {
             ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
         }
+
         articleService.deleteArticle(id);
 
         return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제하였습니다.", id), id);
